requirements.txt
pydantic-ai>=0.0.14
httpx>=0.25.0
anyio>=4.0.0
python-dotenv>=1.0.0 